r.stream.distance: Change the output type from FCELL to DCELL for a higher precision (ticket https://trac.osgeo.org/grass/ticket/2359)
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@61206 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ raster/r.stream.distance/local_vars.h
@@ -22,8 +22,8 @@ typedef struct {
 
 typedef struct {
 	int r,c;
-  float cur_dist;
-  float target_elev;
+  double cur_dist;
+  double target_elev;
 } POINT;
 
 GLOBAL int nextr[9];
